move SerializedProgram into its own file (to fixup type annotations)
